backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic
pydantic[email]
fastapi-jwt-auth
pydantic_settings